Move the mutation rate from class to algorithm.
If there was a specific reason for having the mutation rate in the
params class
let me know but I've moved it to better reflect that it's a property of
the algorithm. This is more in line with the fact that the
representation we write to file does not hold the mutation rate for
example.

This doesn't stop having algorithms that would have different
mutation rates for different parameter instances (it's just that the
algorithm would have to keep track of it).