Implement genetic algorithm for gambler (#48)
* Implement GA for Gambler archetype

* tweak mutation

* add tests for gambler's ga methods

* fix mutation and adjust tests

* add check to keep genes in bounds 0, 1

* add integration test

* Implement GA for Gambler archetype

* tweak mutation

* add tests for gambler's ga methods

* fix mutation and adjust tests

* add check to keep genes in bounds 0, 1

* add integration test

* tweak old tests

* test that player instance is created correctly

* test instance with vector

Tests that an instance of the Gambler is created correctly
when a vector is passed. That's when gambler works with
pso

* fix test on pso

This this the final tweak. Now the tests is correct because we
playing against 5 cooperators and the payoff we expect is around
5.

This was breaking previously because the player instance was not
created properly.

* typo

* fix mutation based on pr discussion.

Change the mutation to be:

- each gene has the probability to change to a number in [0, 1]

Remove the test about the genes beging between the bounds.

* fix style suggestion

* address minor comments on pr

* fix crossover as suggested on pr

* fix mutation as suggested on pr

* add line at the end of file + pep8